requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
anthropic>=0.50.0
chromadb>=0.5.23
pypdf>=5.0.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
requests>=2.32.3
python-multipart>=0.0.12
python-dotenv>=1.0.1
